feat: splitting components and created alias
