+++ components/Carousel/index.ts
@@ -1,0 +1,3 @@
+import { Carousel } from "./Carousel";
+
+export default Carousel;
