+++ components/Carousel/styles.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components/native";
+
+export const ImageContainer = styled.View`
+  align-items: center;
+  justify-content: center;
+`;
+
+export const Image = styled.Image`
+  width: 320px;
+  height: 170px;
+  border-radius: 10px;
+`;
