+++ components/Carousel/types.ts
@@ -1,4 +1,4 @@
-export type Image = {
+export type ImageType = {
   id: string;
   uri: string;
 };
