+++ components/ShowCaption/data.ts
@@ -1,0 +1,6 @@
+export const data = {
+  messages: {
+    copied: "Caption copied to the clipboard!",
+  },
+  placeholder: "Your caption will appear here...",
+};
