+++ components/ShowCaption/index.ts
@@ -1,0 +1,3 @@
+import { ShowCaption } from "./ShowCaption";
+
+export default ShowCaption;
